% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{blrmStats}
\alias{blrmStats}
\title{blrmStats}
\usage{
blrmStats(fit, ns = 400, cint = 0.95, pl = FALSE)
}
\arguments{
\item{fit}{an object produced by \code{blrm}}

\item{ns}{number of posterior draws to use in the calculations (default is 400)}

\item{cint}{credible interval probability (default is 0.95)}

\item{pl}{set to \code{TRUE} to plot the posterior densities using base graphics}
}
\value{
list of class \code{'blrmStats'} whose most important element is \code{Stats}.  The indexes computed are defined below, with gp, B, EV, and vp computed using the intercept corresponding to the median value of Y.  See \url{http://fharrell.com/post/add-value} for more information.
\describe{
 \item{"Dxy"}{Somers' Dxy rank correlation between predicted and observed.  The concordance probability (c-index; AUROC in the binary Y case) may be obtained from the relationship Dxy=2(c-0.5).}
 \item{"g"}{Gini's mean difference: the average absolute difference over all pairs of linear predictor values}
 \item{"gp"}{Gini's mean difference on the predicted probability scale}
 \item{"B"}{Brier score}
 \item{"EV"}{explained variation}
 \item{"v"}{variance of linear predictor}
 \item{"vp"}{variable of estimated probabilities}
}
}
\description{
Compute Indexes of Predictive Accuracy and Their Uncertainties
}
\details{
For a binary or ordinal logistic regression fit from \code{blrm}, computes several indexes of predictive accuracy along with credible intervals for them.  Optionally plots their posterior densities.
}
\examples{
\dontrun{
  f <- blrm(...)
  blrmStats(f, pl=TRUE)   # print and plot
}
}
\seealso{
[Hmisc::rcorr.cens]
}
\author{
Frank Harrell
}
