% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(x)
}
\arguments{
\item{x}{a vector representing a categorical variable}
}
\value{
\code{x} unchanged
}
\description{
Cluster Function for Random Effects
}
\details{
Used by \code{orm} and \code{blrm} to signal a categorical variable to generate random effects.
}
\author{
Frank Harrell
}
