% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{compareBmods}
\alias{compareBmods}
\title{compareBmods}
\usage{
compareBmods(..., method = "stacking", r_eff_list = NULL)
}
\arguments{
\item{...}{a series of model fits}

\item{method}{see \link[loo:loo_model_weights]{loo::loo_model_weights}}

\item{r_eff_list}{see \link[loo:loo_model_weights]{loo::loo_model_weights}}
}
\value{
a \code{loo::loo_model_weights} object
}
\description{
Compare Bayesian Model Fits
}
\details{
Uses \code{loo::loo_model_weights} to compare a series of models such as those created with \code{blrm}
}
\author{
Frank Harrell
}
