% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{distSym}
\alias{distSym}
\title{distSym}
\usage{
distSym(x, prob = 0.9, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector representing a sample from a continuous distribution}

\item{prob}{quantile interval coverage}

\item{na.rm}{set to \code{TRUE} to remove \code{NA}s before proceeding.}
}
\value{
a scalar with a value of 1.0 indicating symmetry
}
\description{
Distribution Symmetry Measure
}
\details{
From a sample from a distribution computes a symmetry measure.  By default it is the gap between the mean and the 0.95 quantile divided by the gap between the 0.05 quantile and the mean.
}
\author{
Frank Harrell
}
