% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{fitIf}
\alias{fitIf}
\title{fitIf}
\usage{
fitIf(w)
}
\arguments{
\item{w}{an expression of the form \code{fitname <- fitting_function(...)}}
}
\description{
fitIf
}
\details{
By calling \code{fitIf(fitname <- fitting_function(...))} \code{fitIf} determines whether the fit object has already been created under the file named \code{"fitname.rds"} in the current working directory, and if so loads it.  It also looks for a file with a name such as \code{"fitname-rstan.rds"} and if it exists laods it and adds that object as the \code{rstan} part of the model fit, for \code{rstan} diagnostics.  If the \code{"fitname.rds"} does not exist, runs the model fit, strips off the large \code{rstan} object, stores the short fit object in \code{"fitname.rds"} and stores the full fit object in R object \code{fitname} in the global environment.  To force a re-fit, remove the \code{.rds} object.
}
\author{
Frank Harrell
}
