% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{HPDint}
\alias{HPDint}
\title{HPDint}
\usage{
HPDint(x, prob = 0.95)
}
\arguments{
\item{x}{a vector of posterior draws}

\item{prob}{desired probability coverage}
}
\value{
a 2-vector with elements \code{Lower} and \code{Upper}
}
\description{
Highest Posterior Density Interval
}
\details{
Adapts code from \code{coda::HPDinterval} to compute a highest posterior density interval from posterior samples for a single parameter.  Quoting from the \code{coda} help file, for each parameter the interval is constructed from the empirical cdf of the sample as the shortest interval  for  which  the  difference  in  the  ecdf  values  of  the  endpoints  is  the  nominal  probability.  Assuming that the distribution is not severely multimodal, this is the HPD interval.
}
\author{
Douglas Bates and Frank Harrell
}
