% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocens.r
\name{is.na.Ocens}
\alias{is.na.Ocens}
\title{is.na Method for Ocens Objects}
\usage{
\method{is.na}{Ocens}(x)
}
\arguments{
\item{x}{an object created by \code{Ocens}}
}
\value{
a logical vector whose length is the number of rows in \code{x}, with \code{TRUE} designating observations having one or both columns of \code{x} equal to \code{NA}
}
\description{
is.na Method for Ocens Objects
}
\examples{
Y <- Ocens(c(1, 2, NA, 4))
Y
is.na(Y)
}
