% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{mix_re}
\alias{mix_re}
\title{mix_re}
\usage{
mix_re(mre)
}
\arguments{
\item{mre}{mixing proportion for dual sigmas for random effects with \code{\link[=orm]{orm()}}.  These are typically a function of time in a Markov model, scaled to be between 0 and 1 inclusive.  When \code{mre} is zero the multiplier of normal(0,1) random effects is the first sigma, and when \code{mre} is 1.0 the multiplier is the second sigma.  Otherwise it is a weighted combination of sigmas.  The first sigma is constrained to be positive.}
}
\value{
\code{mre} unchanged
}
\description{
Mixing Function for Dual Random Effect Sigmas
}
\details{
Used by \code{orm} to specify weights for mixing two scaled normal(0,1) random effects.  The main purpose of \code{mre} is to get a more realistic correlation structure in the raw data space when fitting Markov ordinal models with lags in the raw Y values.  What works well in practice is to set \code{mre} to 0.0 for an observation measured at the earliest follow-up time, and \code{mre} to \code{1.0} when the observation is taken at any time after this initial visit time.
}
\examples{
# mix_re(ifelse(t == 1, 0, 1))
}
\author{
Frank Harrell
}
