% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocens.r
\name{Ocens2Surv}
\alias{Ocens2Surv}
\title{Ocens2Surv}
\usage{
Ocens2Surv(Y)
}
\arguments{
\item{Y}{an \code{Ocens} object}
}
\value{
a \code{Surv} object
}
\description{
Converts an \code{Ocens} object to the simplest \code{Surv} object that works for the types of censoring that are present in the data.
}
\examples{
Y <- Ocens(1:3, c(1, Inf, 3))
Ocens2Surv(Y)
}
