% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{plot.PostF}
\alias{plot.PostF}
\title{plot.PostF}
\usage{
\method{plot}{PostF}(
  x,
  ...,
  cint = 0.95,
  label = NULL,
  type = c("linetype", "facet"),
  ltitle = ""
)
}
\arguments{
\item{x}{result of running a function created by \code{PostF}}

\item{...}{other results created by such functions}

\item{cint}{interval probability}

\item{label}{x-axis label if not the expression originally evaluated.  When more than one result is plotted, \code{label} is a vector of character strings, one for each result.}

\item{type}{when plotting more than one result specifies whether to make one plot distinguishing results by line type, or whether to make separate panels}

\item{ltitle}{used of \code{type='linetype'} to specify name of legend for the line types}
}
\value{
\code{ggplot2} object
}
\description{
Plot Posterior Density of \code{PostF}
}
\details{
Computes highest posterior density and posterior mean and median as vertical lines, and plots these on the density function.  You can transform the posterior draws while plotting.
}
\author{
Frank Harrell
}
