% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexVar.r
\name{plot.rexVar}
\alias{plot.rexVar}
\title{plot.rexVar}
\usage{
\method{plot}{rexVar}(
  x,
  xlab = "Relative Explained Variation",
  xlim = NULL,
  pch = 16,
  sort = c("descending", "ascending", "none"),
  margin = FALSE,
  height = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{x}{a vector or matrix created by \code{rexVar}}

\item{xlab}{x-axis label}

\item{xlim}{x-axis limits; defaults to range of all values (limits and point estimates)}

\item{pch}{plotting symbol for dot}

\item{sort}{defaults to sorted predictors in descending order of relative explained variable.  Can set to \code{ascending} or \code{none}.}

\item{margin}{set to \code{TRUE} to show the REV values in the right margin if using base graphics}

\item{height}{optional height in pixels for \code{plotly} graph}

\item{width}{likewise optional width}

\item{...}{arguments passed to \code{dotchart2} or \code{dotchartpl}}
}
\value{
\code{plotly} graphics object if using \code{plotly}
}
\description{
Plot rexVar Result
}
\details{
Makes a dot chart displaying the results of \code{rexVar}.  Base graphics are used unless \code{options(grType='plotly')} is in effect, in which case a \code{plotly} graphic is produced with hovertext
}
\author{
Frank Harrell
}
