% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{plot.rmsb}
\alias{plot.rmsb}
\title{plot.rmsb}
\usage{
\method{plot}{rmsb}(x, which = NULL, nrow = NULL, ncol = NULL, cint = 0.95, ...)
}
\arguments{
\item{x}{an \code{rms} Bayesian fit object}

\item{which}{names of parameters to plot, defaulting to all non-intercepts. Can instead be a vector of integers.}

\item{nrow}{number of rows of plots}

\item{ncol}{number of columns of plots}

\item{cint}{probability for credible interval}

\item{...}{passed to \code{ggplot2::geom_density}}
}
\value{
\code{ggplot2} object
}
\description{
Plot Posterior Densities and Summaries
}
\details{
For an \code{rms} Bayesian fit object, plots posterior densities for selected parameters along with posterior mode, mean, median, and credible interval
}
\author{
Frank Harrell
}
