% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{predict.blrm}
\alias{predict.blrm}
\title{predict.blrm}
\usage{
\method{predict}{blrm}(
  object,
  ...,
  type = c("lp", "fitted", "fitted.ind", "mean", "x", "data.frame", "terms", "cterms",
    "ccterms", "adjto", "adjto.data.frame", "model.frame"),
  se.fit = FALSE,
  codes = FALSE,
  posterior.summary = c("mean", "median", "mode")
)
}
\arguments{
\item{object, ..., type, se.fit, codes}{see \link{predict.lrm}}

\item{posterior.summary}{set to \code{'median'} or \code{'mode'} to use posterior median/mode instead of mean}
}
\value{
a data frame,  matrix, or vector
}
\description{
Make predictions from a \code{blrm} fit
}
\details{
Predict method for \code{blrm} objects
}
\examples{
\dontrun{
  f <- blrm(...)
  predict(f, newdata, posterior.summary='median')
}
}
\seealso{
\link{predict.lrm}
}
\author{
Frank Harrell
}
