% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{print.blrmStats}
\alias{print.blrmStats}
\title{print.blrmStats}
\usage{
\method{print}{blrmStats}(x, dec = 3, ...)
}
\arguments{
\item{x}{an object produced by \code{blrmStats}}

\item{dec}{number of digits to round indexes}

\item{...}{ignored}
}
\description{
Print Details for \code{blrmStats} Predictive Accuracy Measures
}
\details{
Prints results of \code{blrmStats} with brief explanations
}
\examples{
\dontrun{
  f <- blrm(...)
  s <- blrmStats(...)
  s    # print with defaults
  print(s, dec=4)
}
}
\author{
Frank Harrell
}
