% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactPO.r
\name{print.impactPO}
\alias{print.impactPO}
\title{Print Result from impactPO}
\usage{
\method{print}{impactPO}(x, estimates = nrow(x$estimates) < 16, ...)
}
\arguments{
\item{x}{an object created by \code{impactPO}}

\item{estimates}{set to \code{FALSE} to suppess printing estimated category probabilities.  Defaults to \code{TRUE} when the number of rows < 16.}

\item{...}{ignored}
}
\description{
Prints statistical summaries and optionally predicted values computed by \code{impactPO}, transposing statistical summaries for easy reading
}
\author{
Frank Harrell
}
