% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocens.r
\name{print.Ocens}
\alias{print.Ocens}
\title{print Method for Ocens Objects}
\usage{
\method{print}{Ocens}(x, ivalues = FALSE, digits = 5, ...)
}
\arguments{
\item{x}{an object created by \code{Ocens}}

\item{ivalues}{set to \code{TRUE} to print integer codes instead of character levels when original data were factors or character variables}

\item{digits}{number of digits to the right of the decimal place used in rounding original levels when \code{ivalues=FALSE}}

\item{...}{ignored}
}
\value{
nothing
}
\description{
print Method for Ocens Objects
}
\examples{
Y <- Ocens(1:3, c(1, Inf, 3))
Y
print(Y, ivalues=TRUE)  # doesn't change anything since were numeric
}
