% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{print.predict.blrm}
\alias{print.predict.blrm}
\title{print.predict.blrm}
\usage{
\method{print}{predict.blrm}(x, digits = 3, ...)
}
\arguments{
\item{x}{result from \code{predict.blrm}}

\item{digits}{number of digits to round numeric results}

\item{...}{ignored}
}
\description{
Print Predictions for \code{blrm}
}
\details{
Prints the summary portion of the results of \code{predict.blrm}
}
\author{
Frank Harrell
}
