% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexVar.r
\name{print.rexVar}
\alias{print.rexVar}
\title{print.rexVar}
\usage{
\method{print}{rexVar}(x, title = "Relative Explained Variation", digits = 3, ...)
}
\arguments{
\item{x}{a vector or matrix created by \code{rexVar}}

\item{title}{character string which can be set to \code{NULL} or \code{''} to suppress}

\item{digits}{passed to \code{\link[=round]{round()}}}

\item{...}{unused}
}
\value{
invisible
}
\description{
Print rexVar Result
}
\details{
Prints the results of an \code{rexVar} call
}
\author{
Frank Harrell
}
