% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{print.rmsb}
\alias{print.rmsb}
\title{print.rmsb}
\usage{
\method{print}{rmsb}(
  x,
  posterior.summary = c("mean", "median"),
  cint = 0.95,
  dec = 4,
  pr = TRUE,
  ...
)
}
\arguments{
\item{x}{an object created by an \code{rms} Bayesian fitting function}

\item{posterior.summary}{set to \code{'median'} compute posterior median coefficients instead of means}

\item{cint}{credible interval coverage probability (default is 0.95)}

\item{dec}{amount of rounding (digits to the right of the decimal)}

\item{pr}{set to \code{FALSE} to return an unrounded matrix and not print}

\item{...}{ignored}
}
\value{
matrix (rounded if \code{pr=TRUE})
}
\description{
Basic Print for Bayesian Parameter Summary
}
\details{
For a Bayesian regression fit prints the posterior mean, median, SE, credible interval, and symmetry coefficient from the posterior draws.  For a given parameter, the symmetry is the gap between the mean and 0.94 quantile divided by the gap between the 0.05 quantile and the mean.
}
\examples{
\dontrun{
  f <- blrm(...)
  print.rmsb(f)
}
}
\author{
Frank Harrell
}
