% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMI.r
\name{prmiInfo}
\alias{prmiInfo}
\title{prmiInfo}
\usage{
prmiInfo(x)
}
\arguments{
\item{x}{an object created by \code{processMI(..., 'anova')}}
}
\value{
nothing
}
\description{
Print Information About Impact of Imputation
}
\details{
For the results of \code{processMI.fit.mult.impute} prints or renders a formatted table (html, LaTeX, or Typst, according to \code{options(prType=)}) summarizing various correction factors related to missing data multiple imputation.
}
\examples{
\dontrun{
a <- aregImpute(...)
f <- fit.mult.impute(...)
v <- processMI(f, 'anova')
prmiInfo(v)
}
}
\author{
Frank Harrell
}
