% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMI.r
\name{processMI}
\alias{processMI}
\title{processMI}
\usage{
processMI(object, ...)
}
\arguments{
\item{object}{a fit object created by \code{\link[Hmisc:transcan]{Hmisc::fit.mult.impute()}}}

\item{...}{ignored}
}
\value{
an object that resembles something created by a single fit without multiple imputation
}
\description{
Process Special Multiple Imputation Output
}
\details{
Processes lists that have one element per imputation
}
\seealso{
\code{\link[=processMI.fit.mult.impute]{processMI.fit.mult.impute()}}
}
\author{
Frank Harrell
}
