% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glm.r
\name{residuals.Glm}
\alias{residuals.Glm}
\title{residuals.Glm}
\usage{
\method{residuals}{Glm}(object, type, ...)
}
\arguments{
\item{object}{a fit object produced by `Glm`}

\item{type}{either `'score'` or a `type` accepted by `residuals.glm`}

\item{...}{ignored}
}
\value{
a vector or matrix
}
\description{
Residuals for `Glm`
}
\details{
This function mainly passes through to `residuals.glm` but for `type='score'` computes the matrix of score residuals using code modified from `sandwich::estfun.glm`.
}
\author{
Frank Harrell
}
