% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms-package.R
\docType{package}
\name{rms-package}
\alias{rms}
\alias{rms-package}
\title{rms: CMHC Rental Market Survey Data}
\description{
`rms` packages Rental Market Survey RMS data from CMHC for convenient access in R. Currently supports Rental Market Report data from 2018 to 2023.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/context-dependent/rms}
  \item \url{https://context-dependent.github.io/rms/}
  \item Report bugs at \url{https://github.com/context-dependent/rms/issues}
}

}
\author{
\strong{Maintainer}: Thomas McManus \email{tmcmanus@blueprint-ade.ca}

}
\keyword{internal}
