% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rsimms}
\alias{rsimms}
\title{Simulate multi-state survival data with weibull}
\usage{
rsimms(beta01 = 0, covs, maxt = 50, beta02 = 0, beta12 = 0)
}
\description{
Simulate multi-state survival data with weibull
}
