% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rsimmswei}
\alias{rsimmswei}
\title{Simulate multi-state survival data with weibull}
\usage{
rsimmswei(lambdas01, gammas01, beta01 = 0, covs, maxt = 50, lambdas02,
  gammas02, beta02 = 0, lambdas12, gammas12, beta12 = 0)
}
\description{
Simulate multi-state survival data with weibull
}
