% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackMI.r
\name{stackMI}
\alias{stackMI}
\title{Bayesian Model Fitting and Stacking for Multiple Imputation}
\usage{
stackMI(
  formula,
  fitter,
  xtrans,
  data,
  n.impute = xtrans$n.impute,
  dtrans,
  derived,
  subset,
  refresh = 0,
  progress = if (refresh > 0) "stan-progress.txt" else "",
  ...
)
}
\arguments{
\item{formula}{a model formula}

\item{fitter}{a Bayesian fitter}

\item{xtrans}{an object created by \code{transcan}, \code{\link{aregImpute}}, or \code{\link[mice]{mice}}}

\item{data}{data frame}

\item{n.impute}{number of imputations to run, default is the number saved in \code{xtrans}}

\item{dtrans}{see \code{Hmisc::fit.mult.impute}}

\item{derived}{see \code{Hmisc::fit.mult.impute}}

\item{subset}{an integer or logical vector specifying the subset of observations to fit}

\item{refresh}{see [rstan::sampling].  The default is 0, indicating that no progress notes are output.  If \code{refresh > 0} and \code{progress} is not \code{''}, progress output will be appended to file \code{progress}.  The default file name is \code{'stan-progress.txt'}.}

\item{progress}{see \code{refresh}.  Defaults to \code{''} if \code{refresh = 0}.  Note: If running interactively but not under RStudio, \code{rstan} will open a browser window for monitoring progress.}

\item{...}{arguments passed to \code{fitter}}
}
\value{
an \code{rms} fit object with expanded posterior draws and diagnostics
}
\description{
Runs an \code{rms} package Bayesian fitting function such as \code{blrm} separately for each completed dataset given a multiple imputation result such as one produced by \code{Hmisc::aregImpute}.  Stacks the posterior draws and diagnostics across all imputations, and computes parameter summaries on the stacked posterior draws.
stackMI
}
\author{
Frank Harrell
}
