% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanCompile}
\alias{stanCompile}
\title{Compile Stan Code}
\usage{
stanCompile(repo = "https://raw.githubusercontent.com/harrelfe/stan/master")
}
\arguments{
\item{repo}{URL to online source file base}
}
\description{
Retrieves Stan code files from the github repository, compiles then with \code{rstan::stan_model}, and stores the compiled code in a central placed defined by the user with e.g. \code{options(stancompiled='~/R/stan')}.
stanCompile
}
\examples{
  \dontrun{
  options(stancompiled='~/R/stan')    # need this always
  stanCompile()    # do this once per computer to compile centrally
  }
}
\author{
Frank Harrell
}
