% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanDx}
\alias{stanDx}
\title{stanDx}
\usage{
stanDx(object)
}
\arguments{
\item{object}{an object created by an \code{rms} package Bayesian fitting function such as \code{blrm} or \code{stackMI}}
}
\value{
matrix suitable for printing
}
\description{
Print Stan Diagnostics
}
\details{
Retrieves the effect samples sizes and Rhats computed after a fitting function ran \code{rstan}, and prepares it for printing.  If the fit was created by \code{stackImpute}, the diagnostics for all imputations are printed (separately).
}
\examples{
\dontrun{
  f <- blrm(...)
  stanDx(f)
}
}
\author{
Frank Harrell
}
