% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanDxplot}
\alias{stanDxplot}
\title{stanDxplot}
\usage{
stanDxplot(x, which = x$betas, ...)
}
\arguments{
\item{x}{an \code{rms} Bayesian fit object}

\item{which}{names of parameters to plot, defaulting to all non-intercepts}

\item{...}{passed to \code{rstan::traceplot}}
}
\value{
\code{ggplot2} object
}
\description{
Diagnostic Trace Plots
}
\details{
For an \code{rms} Bayesian fit object, calls the \code{rstan} \code{traceplot} function on the \code{rstan} object inside the \code{rmsb} object, to check properties of posterior sampling
}
\author{
Frank Harrell
}
