% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanGet}
\alias{stanGet}
\title{stanGet}
\usage{
stanGet(object)
}
\arguments{
\item{object}{an objected created by an \code{rms} package Bayesian fitting function}
}
\value{
the object created by \code{rstan::sampling}
}
\description{
Get Stan Output
}
\details{
Extracts the object created by \code{rstan::sampling} so that standard Stan diagnostics can be run from it
}
\examples{
\dontrun{
  f <- blrm(...)
  s <- stanGet(f)
}
}
\author{
Frank Harrell
}
