% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanQr}
\alias{stanQr}
\title{stanQr}
\usage{
stanQr(X, center = TRUE)
}
\arguments{
\item{X}{a numeric matrix}

\item{center}{set to \code{FALSE} to not center columns of \code{X} first}
}
\value{
list with elements \code{Xqr, R}
}
\description{
Stan Thin QR Decomposition
}
\details{
Runs a matrix through the Stan thin QR decomposition as used in \code{rms} Bayesian fitting functions and returns the transformed matrix \code{Xqr} and the reverse transforming matrix \code{R}.  If columns of the input matrix \code{X} are centered the QR transformed matrix will be orthogonal.  This is helpful in understanding the transformation and in scaling prior distributions on the transformed scale.
}
\examples{
\dontrun{
  x <- 1 : 10
  X <- cbind(x, x^2)
  w <- stanQr(X)
  w
  Xqr <- w$Xqr
  plot(X[, 1], Xqr[, 1])
  plot(X[, 1], Xqr[, 2])
  cov(X)
  cov(Xqr)
}
}
\author{
Frank Harrell
}
