% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocens.r
\name{[.Ocens}
\alias{[.Ocens}
\title{Ocens}
\usage{
\method{[}{Ocens}(x, ..., drop)
}
\arguments{
\item{x}{an \code{Ocens} object}

\item{...}{the usual rows and columns specifiers}

\item{drop}{set to \code{FALSE} to not drop unneeded dimensions}
}
\value{
new \code{Ocens} object or by default an unclassed vector if only one column of \code{x} is being kept
}
\description{
Subset Method for \code{Ocens} Objects
}
\details{
Subsets an \code{Ocens} object, preserving its special attributes.  Attributes are not updated.  In the future such updating should be implemented.
}
\author{
Frank Harrell
}
