% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xcontrast.r
\name{Xcontrast}
\alias{Xcontrast}
\title{Xcontrast}
\usage{
Xcontrast(
  fit,
  a,
  b = NULL,
  a2 = NULL,
  b2 = NULL,
  ycut = NULL,
  weights = "equal",
  expand = TRUE
)
}
\arguments{
\item{fit}{an `rms` or `rmsb` fit object, not necessarily complete}

\item{a}{see [rms::contrast.rms()]}

\item{b}{see [rms::contrast.rms()]}

\item{a2}{see [rms::contrast.rms()]}

\item{b2}{see [rms::contrast.rms()]}

\item{ycut}{see [rms::contrast.rms()]}

\item{weights}{see [rms::contrast.rms()]}

\item{expand}{see [rms::contrast.rms()]}
}
\value{
numeric matrix
}
\description{
Produce Design Matrices for Contrasts
}
\details{
This is a simpler version of `contrast.rms` that creates design matrices or differences of them and does not require the fit object to be complete (i.e., to have coefficients).  This is used for the `pcontrast` option in [rmsb::blrm()].
}
\author{
Frank Harrell
}
