% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{fitLoad}
\alias{fitLoad}
\title{fitLoad}
\usage{
fitLoad(fit)
}
\arguments{
\item{name}{R object name (unquoted) that was passed to \code{fitSave}}
}
\value{
an object, using an rms package fit object
}
\description{
Load a Saved rms Object
}
\details{
Adds a suffix of \code{.rds} to the name of the argument and calls \code{readRDS} to read the stored object, returning it to the user
}
\author{
Frank Harrell
}
