% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactPO.r
\name{impactPO}
\alias{impactPO}
\title{Impact of Proportional Odds Assumpton
Checks the impact of the proportional odds assumption by comparing predicted cell probabilities from a PO model with those from a multinomial or partial proportional odds logistic model that relax assumptions.  For a given model formula, fits the model with both \code{lrm} and either \code{nnet::multinom} or \code{VGAM::vglm} or both, and obtains predicted cell probabilities for the PO and relaxed models on the \code{newdata} data frame.}
\usage{
impactPO(
  formula,
  relax = if (missing(nonpo)) "multinomial" else "both",
  nonpo,
  newdata,
  ...
)
}
\arguments{
\item{formula}{a model formula.  To work properly with \code{multinom} or \code{vglm} the terms should have completely specified knot locations if a spline function is being used.}

\item{relax}{defaults to \code{"both"} if \code{nonpo} is given, resulting in fitting two relaxed models.  Set \code{relax} to \code{"multinomial"} or \code{"ppo"} to fit only one relaxed model.  The multinomial model does not assume PO for any predictor.}

\item{nonpo}{a formula with no left hand side variable, specifying the variable or variables for which PO is not assumed.  Specifying \code{nonpo} results in a relaxed fit that is a partial PO model fitted with \code{VGAM::vglm}.}

\item{newdata}{a data frame or data table with one row per covariate setting for which predictions are to be made}

\item{...}{other parameters to pass to \code{lrm} and \code{multinom} such as \verb{data=}.}
}
\value{
an \code{impactPO} object which is a list with elements \code{estimates} and \code{stats}.  \code{data} is a data frame containing the variables and values in \code{newdata} in a tall and thin format with additional variable \code{method} ("PO", "multinomial", "PPO"), \code{y} (current level of the dependent variable), and \code{Probability} (predicted cell probability for covariate values and value of \code{y} in the current row).  \code{stats} is a data frame containing \code{Deviance} the model deviance, \code{d.f.} the total number of parameters counting intercepts, \code{AIC}, \code{p} the number of regression coefficients, \verb{LR chi^2} the likelihood ratio chi-square statistic for testing the predictors, \code{LR - p} a chance-corrected LR chi-square, \code{R2} the Nagelkerke pseudo R-squared, and \verb{R2 adj} a chance-corrected pseudo R-squared using \code{LR - p}.
}
\description{
Impact of Proportional Odds Assumpton
Checks the impact of the proportional odds assumption by comparing predicted cell probabilities from a PO model with those from a multinomial or partial proportional odds logistic model that relax assumptions.  For a given model formula, fits the model with both \code{lrm} and either \code{nnet::multinom} or \code{VGAM::vglm} or both, and obtains predicted cell probabilities for the PO and relaxed models on the \code{newdata} data frame.
}
\examples{

\dontrun{
set.seed(1)
age <- rnorm(500, 50, 10)
sex <- sample(c('female', 'male'), 500, TRUE)
y   <- sample(0:4, 500, TRUE)
d   <- expand.grid(age=50, sex=c('female', 'male'))
w   <- impactPO(y ~ age + sex, nonpo = ~ sex, newdata=d)
w
# Note that PO model is a better model than multinomial (lower AIC)
# since multinomial model's improvement in fit is low in comparison
# with number of additional parameters estimated.  Same for PO model
# in comparison with partial PO model.

# Reverse levels of y so stacked bars have higher y located higher
revo <- function(z) {
  z <- as.factor(z)
  factor(z, levels=rev(levels(as.factor(z))))
}

ggplot(w$estimates, aes(x=method, y=Probability, fill=revo(y))) +
  facet_wrap(~ sex) + geom_col() +
  xlab('') + guides(fill=guide_legend(title=''))

# Now vary 2 predictors

d <- expand.grid(sex=c('female', 'male'), age=c(40, 60))
w <- impactPO(y ~ age + sex, nonpo = ~ sex, newdata=d)
w
ggplot(w$estimates, aes(x=method, y=Probability, fill=revo(y))) +
  facet_grid(age ~ sex) + geom_col() +
 xlab('') + guides(fill=guide_legend(title=''))
}
}
\author{
Frank Harrell \href{mailto:fh@fharrell.com}{fh@fharrell.com}
}
