% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactPO.r
\name{impactPO}
\alias{impactPO}
\title{Impact of Proportional Odds Assumpton
Checks the impact of the proportional odds assumption by comparing predicted cell probabilities from a PO model with those from a multinomial logistic model.  For a given model formula, fits the model with both \code{lrm} and \code{nnet::multinom} and obtains predicted cell probabilities for both on the \code{newdata} data frame.}
\usage{
impactPO(formula, newdata, ...)
}
\arguments{
\item{formula}{a model formula.  To work properly with \code{multinom} the terms should have completely specified knot locations if a spline function is being used.}

\item{newdata}{a data frame or data table with one row per covariate setting for which predictions are to be made}

\item{...}{other parameters to pass to \code{lrm} and \code{multinom} such as \verb{data=}.}
}
\value{
an \code{impactPO} object which is a list with elements \code{estimates} and \code{stats}.  \code{data} is a data frame containing the variables and values in \code{newdata} in a tall and thin format with additional variable \code{method} ("PO" or "multinomial"), \code{y} (current level of the dependent variable), and \code{Probability} (predicted cell probability for covariate values and value of \code{y} in the current row).
}
\description{
Impact of Proportional Odds Assumpton
Checks the impact of the proportional odds assumption by comparing predicted cell probabilities from a PO model with those from a multinomial logistic model.  For a given model formula, fits the model with both \code{lrm} and \code{nnet::multinom} and obtains predicted cell probabilities for both on the \code{newdata} data frame.
}
\examples{

set.seed(1)
age <- rnorm(500, 50, 10)
sex <- sample(c('female', 'male'), 500, TRUE)
y   <- sample(0:4, 500, TRUE)
d   <- expand.grid(age=50, sex=c('female', 'male'))
w   <- impactPO(y ~ age + sex, d)
w
# Note that PO model is a better model than multinomial (lower AIC)
# since multinomial model's improvement in fit is low in comparison
# with number of additional parameters estimated

# Reverse levels of y so stacked bars have higher y located higher
revo <- function(z) {
  z <- as.factor(z)
  factor(z, levels=rev(levels(as.factor(z))))
}

ggplot(w$estimates, aes(x=method, y=Probability, fill=revo(y))) +
  facet_wrap(~ sex) + geom_col() +
  xlab('') + guides(fill=guide_legend(title=''))

d <- expand.grid(sex=c('female', 'male'), age=c(40, 60))
w <- impactPO(y ~ age + sex, d)
w
ggplot(w$estimates, aes(x=method, y=Probability, fill=revo(y))) +
  facet_grid(age ~ sex) + geom_col() +
 xlab('') + guides(fill=guide_legend(title=''))
}
\author{
Frank Harrell \href{mailto:fh@fharrell.com}{fh@fharrell.com}
}
