\name{rms-internal}
\title{Internal rms functions}
\alias{annotateAnova}
\alias{coxphFit}
\alias{lm.pfit}
\alias{ols.influence}
\alias{plotmathAnova}
\alias{prType}
\alias{survreg.auxinfo}
\alias{val.probg}
\description{Internal rms functions}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
