% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{stanCompile}
\alias{stanCompile}
\title{Compile Stan Code}
\usage{
stanCompile(
  mods = c("lrm", "lrmc", "lrmppo", "lrmcppo", "lrmcar1"),
  repo = "https://raw.githubusercontent.com/harrelfe/stan/master"
)
}
\arguments{
\item{mods}{character vector of base names of \code{.stan} model files to compile.  Default is to compile all those currently in use in \code{rms}}

\item{repo}{URL to online source file base}
}
\description{
Retrieves Stan code files from the github repository, compiles then with \code{rstan::stan_model}, and stores the compiled code in a central placed defined by the user with e.g. \code{options(stancompiled='~/R/stan')}.
stanCompile
}
\examples{
  \dontrun{
  options(stancompiled='~/R/stan')    # need this always
  stanCompile()    # do this once per computer to compile centrally
  }
}
\author{
Frank Harrell
}
