% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ocens.r
\name{[.Ocens}
\alias{[.Ocens}
\title{Ocens}
\usage{
\method{[}{Ocens}(x, rows = 1:d[1], cols = 1:d[2], ...)
}
\arguments{
\item{x}{an \code{Ocens} object}

\item{rows}{logical or integer vector}

\item{cols}{logical or integer vector}

\item{...}{ignored}
}
\value{
new \code{Ocens} object
}
\description{
Subset Method for \code{Ocens} Objects
}
\details{
Subsets an \code{Ocens} object, preserving its special attributes.  Attributes are not updated.  In the future such updating should be implemented.
}
\author{
Frank Harrell
}
